Address code comments

+++ pycloudlib/openstack/instance.py
@@ -1,8 +1,13 @@
 """Openstack instance type."""
 import time
+from itertools import chain
 
 import openstack
-from openstack.exceptions import BadRequestException, ConflictException
+from openstack.exceptions import (
+    BadRequestException,
+    ConflictException,
+    ResourceNotFound,
+)
 
 from pycloudlib.instance import BaseInstance
 
@@ -20,7 +25,6 @@ class OpenstackInstance(BaseInstance):
             instance_id: The instance id representing the cloud instance
             connection: The connection used to create this instance.
                 If None, connection will be created.
-
         """
         super().__init__(key_pair)
 
@@ -30,11 +34,24 @@ class OpenstackInstance(BaseInstance):
 
         self.server = self.conn.compute.get_server(instance_id)
 
-        self.floating_ip = self.conn.create_floating_ip(
-            wait=True,
-        )
+        self.delete_floating_ip = False
+        self.floating_ip = self._get_existing_floating_ip()
+        if self.floating_ip is None:
+            self._create_and_attach_floating_id()
+            self.delete_floating_ip = True
+
+    def _get_existing_floating_ip(self):
+        server_addresses = chain(*self.server.addresses.values())
+        server_ips = [addr['addr'] for addr in server_addresses]
+        for floating_ip in self.conn.network.ips():
+            if floating_ip['floating_ip_address'] in server_ips:
+                return floating_ip
+        return None
+
+    def _create_and_attach_floating_id(self):
+        self.floating_ip = self.conn.create_floating_ip(wait=True)
         tries = 30
-        while tries:
+        for _ in range(tries):
             try:
                 self.conn.compute.add_floating_ip_to_server(
                     self.server,
@@ -43,7 +60,6 @@ class OpenstackInstance(BaseInstance):
                 break
             except BadRequestException as e:
                 if 'Instance network is not ready yet' in str(e):
-                    tries -= 1
                     time.sleep(1)
                     continue
                 raise e
@@ -76,8 +92,14 @@ class OpenstackInstance(BaseInstance):
         Args:
             wait: wait for instance to be deleted
         """
-        self.conn.compute.delete_server(self.server.id)
-        self.conn.delete_floating_ip(self.floating_ip.id)
+        try:
+            self.conn.compute.delete_server(self.server.id)
+        finally:
+            if self.delete_floating_ip:
+                self.conn.delete_floating_ip(self.floating_ip.id)
+        if wait:
+            self.wait_for_delete()
+
 
     def restart(self, wait=True, **kwargs):
         """Restart the instance.
@@ -107,6 +129,7 @@ class OpenstackInstance(BaseInstance):
         try:
             self.conn.compute.start_server(self.server)
         except ConflictException as e:
+            # We can get an exception here if the instance is already started
             if 'while it is in vm_state active' in str(e):
                 return
         if wait:
@@ -118,7 +141,11 @@ class OpenstackInstance(BaseInstance):
 
     def wait_for_delete(self):
         """Wait for instance to be deleted."""
-        self.conn.compute.wait_for_server(self.server, status='DELETED')
+        try:
+            self.conn.compute.wait_for_server(self.server, status='DELETED')
+        except ResourceNotFound:
+            # We can 404 here is instance is already deleted
+            pass
 
     def wait_for_stop(self):
         """Wait for instance stop."""
